Add sample data to create segmentation demo
Since most people won't have any compatible dicom data
available, we add the ability to grab a sample dataset
to use as the basis for painting.

For now the 8 slices of the MRHead are served from
amazon s3 to allow CORS access to data that is not
in the respository.