Update structured report example
Now it makes reports that are more or less technically correct
with only a few exceptions.

They are correct in the sense that dsrdump from DCMTK can correctly
intepret the results with only one warning (see issue #3).

They are also correct in the sense that they pass David Clunie's
dciodvfy tool except for an error also related to issue #3 and
that the version of dciodvfy I have doesn't know about
attribute 0008,0118 (MappingResourceUID) which must have come
into the standard after the version of dciodvfy that I have.

Otherwise the generated reports have not been widely tested.