Merge branch 'add-packaging'
Adds ES2015 module support

Use these changes with 'rollup -c' to
create a dcmjs.js and dsmjs.js.map in
the build directory.

Use the result with a script tag
in the examples.  DCMJS is the
new overall module name in browser.

Adds import and export support
packaging
and clarify relationship
across files.