fix(naturalizing): Fix the exception on naturalize twice (#237)
The problem with naturalizing twice, was that the change affected the
original data object as it was originally.  The exception was caused by
adding two defineProperty on the same object, the subsequent time.
The fix is twofold, first to avoid any changes to the underlying object
by copying the data to a new list object.  The second change is to allow
the defineProperty to be redefinable, so even if it got applied to the
same object a second time, there wouldn't be a problem.