Fix(naturalize): large objects #263 (#267)
* fix(dcmjs):Use a proxy to fix #263

The issue was that there was an accessor being added for a very large
integer (0x60030010), causing the array to be that size, but with a null
value for everything up to that size.  Changed to use a proxy value
instead, which allows differentiation by string type, and by whether the
underlying object has the original value.

* fix(dcmjs):Add a set on addAccessor results unit test.