feat(doubleClick): Add Double click detection (#375) (#382)
* fix(mouseDownListener): Double click detection (#375)

Added a timeout in mouseDownListener to determine if a double click occurs.
If a double click occurs during the timeout,
then any mouse down and up events that occurred during the timeout are suppressed.
If the timeout expires, then the first mouse down and up events during the timeout
are fired and any browser double click event that might occur is suppressed/ignored.
Double clicks that occur during tool interaction are suppressed
Double click event capture phase listeners are used to more reliably suppress double clicks.

* Fixing syntax error in example-info.json.

* PR feedback: CS3D should not suppress a double click it has detected in
mouseDownListener - even during annotation creation. It is up to the any
listener of the dblclick event to decide how it should be handled.

* PR feedback - using a drag tolerance to help determine if a double click will not occur

* PR updates:

- code comments added and modified for clarification
- the double click listener to ignore double click events is now entirely
encapsulated in the mouseDownListener module
- better named variables in the unit tests
- better conversion of world CPU coordinates to Point3

* Added a TODO for the double click tolerance for touch devices.

* PR feedback - restored adding and removing the double click ignore listener
with other mouse events

* Adding the ignore double click listener back on the element.

* Adding and removing the ignore double click listener in the tools init
and destroy functions (respectively).
Double click drag tolerance is now a projected distance.