refactor: improve code maintainability by extracting rendering logic to separate functions for CPU and GPU
This commit improves the maintainability and reusability of the codebase by separating image rendering function into two separate functions `renderToCanvasCPU` and `renderToCanvasGPU`. The code responsible for rendering Cornerstone images to the canvas relied on an implementation in which we always used the CPU to render the images. However, combining the CPU and GPU rendering paths would have resulted in more complex and error-prone code. Therefore, it was decided to refactor the codebase to improve maintainability by having two separate functions for CPU and GPU. This improves the code by reducing complexity and increasing readability.