feat(streaming-image-volume): add caching for image load object
This commit adds functionality to cache image load object when it
has not already been loaded into the cache. This optimizes the
process of loading images.

refactor(stackToVolumeWithAnnotations): reverse order of image stack

This commit reverses the order of images in the image stack to match
the order in which they were added to the stack.

refactor(stackToVolumeWithAnnotations): use sagittal as default axis

This commit changes the default axis for the stack viewport to sagittal.