feat(rendering): optimize pixel data handling for vtkImageData and other rendering libraries
This commit replaces the "typedArray" property of "StackViewport" with the "pixelArray" property to avoid confusion with TypedArray. We also ensure that the rendering library vtkImageData receives an optimized Uint8Array format for RGB images. Other changes include renaming "getTypedArrayFromMinMax" to "getPixelDataTypeFromMinMax" and returning the pixel data type instead of the TypedArray constructor for better semantic representation. We also ensure that the new function uses the correct numeric range for the different TypedArrays. Finally, in "createImage" and "decodeImageFrame" we apply these changes to improve the performance of rendering with image libraries.