feat(dicomImageLoader): Added support for RGB and 8 bits color image types
This commit adds support for RGB and 8 bits color image types while decoding the image data. The commit also updates the code to convert the pixel data to rgba if decoding is done as RGB. The new code also takes alpha values into account when adding them to the pixel data.