feat(RenderingEngine): Add async rendering to the rendering engine's CPU & GPU pipelines
This commit adds two new functions to the rendering engine's GPU & CPU pipelines: `renderToCanvasGPU` and `renderToCanvasCPU`, both of which are defined in the `utilities` module. These two new functions allow for the rendering of an image to a canvas via the graphics card or the CPU, respectively. They replace the previously existing `renderToCanvas` function which no longer exists. This change introduces the use of promises for asynchronous rendering. In addition, the `renderingEngineCache` now sorts rendering engines based on their `id` property, with engines that start with an underscore `_` coming last in the order.