feat(renderCanvasGPU): use gpu to in the renderToCanvas utility (#586)
* refactor: improve code maintainability by extracting rendering logic to separate functions for CPU and GPU

This commit improves the maintainability and reusability of the codebase by separating image rendering function into two separate functions `renderToCanvasCPU` and `renderToCanvasGPU`. The code responsible for rendering Cornerstone images to the canvas relied on an implementation in which we always used the CPU to render the images. However, combining the CPU and GPU rendering paths would have resulted in more complex and error-prone code. Therefore, it was decided to refactor the codebase to improve maintainability by having two separate functions for CPU and GPU. This improves the code by reducing complexity and increasing readability.

* wip

* wip

* wip

* feat(RenderingEngine): Add async rendering to the rendering engine's CPU & GPU pipelines

This commit adds two new functions to the rendering engine's GPU & CPU pipelines: `renderToCanvasGPU` and `renderToCanvasCPU`, both of which are defined in the `utilities` module. These two new functions allow for the rendering of an image to a canvas via the graphics card or the CPU, respectively. They replace the previously existing `renderToCanvas` function which no longer exists. This change introduces the use of promises for asynchronous rendering. In addition, the `renderingEngineCache` now sorts rendering engines based on their `id` property, with engines that start with an underscore `_` coming last in the order.

* refactor(core/renderingengine): improve viewport resizing

Simplify how viewports are resized by separating VTK-driven viewports and custom rendering viewports. This change will allow for a more efficient and clear codebase.

* wip

* add example for render to canvas

* update api

* apply review comments

* fix removing elements

* update api

* fix nac pt window level

* address review comments