Allow image volume loading to work when SharedArrayBuffer is being used.
For me, scalarData.buffer was a SharedArrayBuffer and not an ArrayBuffer here.
This was causing handleArrayBufferLoad to ignore the successful loading of
pixel data, keeping the volume scalarData all-zero, and eventually causing the
volume to display completely empty even though all images were loaded.

I suppose, at this point, the idea was to check whether there was any
scalarData to load into -- not that the type of buffer was explicitly
ArrayBuffer. By changing the check accordingly, the slice pixel data is being
loaded into the volume scalarData correctly, and the volume is rendered
normally.