feat(VolumeViewport): add rotation property for Volume Viewport (#928)
* feat: add rotation capabilities to volume viewport

* chore: improve method names

* fix: change rotation api to setProperties

* remove protected

* setRotation for volumeViewport

* re-add

* re-add

* update api

* revert on getRotation

* update api

* PR apply comment

* add more comments

* feat(volume): add random rotation feature to
volume API and volumeViewport3D examples

This commit introduces a new feature to the volume API and
volumeViewport3D examples. It adds a button to the toolbar titled "Apply
random rotation". When the button is clicked, the rendering engine
retrieves the volume viewport, applies a random rotation value to the
camera of the viewport, and triggers a re-rendering of the viewport.

These changes improve the user experience by allowing them to apply
random rotations to the rendered volume, enhancing the visual
exploration of the dataset.

Note: The commit also includes minor changes to the Viewport class in
the core package, removing an unnecessary async and promise wait all
function.

---------

Co-authored-by: Pedro Kohler <pedrokohlerbh@gmail.com>
Co-authored-by: Alireza <ar.sedghi@gmail.com>