Initial version of a DICOM custom loader function.
It allows dropping instances or series into an image drop area, and loading the
DICOMs from there.

The current version fails, because it tries to load the DICOMs as browser image
objects. Instead, it should use the DICOM parsing from the dicomImageLoader.
This isn't exposed yet, at the moment, which is the next thing I'm going to
address.